refactor: aplicando princípios de clean code

+++ backend/src/main/groovy/org/jean/linketinder/DAO/CandidateDAO.groovy
@@ -2,12 +2,11 @@ package org.jean.linketinder.DAO
 
 import groovy.sql.Sql
 import org.jean.linketinder.Entities.Candidate
-import org.jean.linketinder.Enum.Skills
 
 class CandidateDAO {
     Sql sql = Sql.newInstance(DBConection.conect()) as Sql
 
-    void Create(Candidate candidate) {
+    void create(Candidate candidate) {
         try {
             sql.execute("INSERT INTO candidates (name, email, cpf, age, state, cep, description) VALUES (?, ?, ?, ?, ?, ?, ?)",
                     [candidate.name, candidate.email, candidate.cpf, candidate.age, candidate.state, candidate.cep, candidate.description])
@@ -19,7 +18,7 @@ class CandidateDAO {
             def id = sql.firstRow("SELECT id FROM candidates WHERE cpf = ?", [candidate.cpf]).id
 
             candidate.skillsList.each { skill ->
-                sql.execute("INSERT INTO candidate_companies (candidate_id, skill_id) VALUES (?, (SELECT skill_id FROM skills WHERE name = ? LIMIT 1))",
+                sql.execute("INSERT INTO candidate_skills (candidate_id, skill_id) VALUES (?, (SELECT skill_id FROM skills WHERE name = ? LIMIT 1))",
                         [id, skill])
             }
 
@@ -29,22 +28,39 @@ class CandidateDAO {
             println("Erro ao adicionar candidate: ${e.message}")
         }
     }
-    List<Candidate> Get() {
-        List<Candidate> candidates = sql.rows("SELECT * FROM candidates") as List<Candidate>
+    List<Candidate> getAll() {
+        List<Candidate> candidates = []
 
-        candidates.each { candidate ->
-
-            def id = sql.firstRow("SELECT id FROM candidates WHERE cpf = ?", [candidate.cpf]).id
-
-            candidate.skillsList = sql.rows("SELECT name FROM skills " +
-                    "INNER JOIN candidate_companies ON skills.skill_id = candidate_companies.skill_id " +
-                    "WHERE candidate_companies.candidate_id = ?", [id]).collect(({ it.name } as Closure<Skills>))
+        sql.eachRow("SELECT id, name, email, cpf, age, state, cep, description FROM candidates") { row ->
+            def skillsForCandidate = getSkillsForCandidate(row.id.toLong()) ?: []
+            Candidate candidate = new Candidate(
+                    row.name,
+                    row.email,
+                    row.cpf,
+                    row.age as Integer,
+                    row.state,
+                    row.cep,
+                    row.description,
+                    skillsForCandidate as List<Skills>
+            )
+            candidates.add(candidate)
         }
-
         return candidates
     }
 
-    void Update(String cpf, Candidate candidate) {
+    private List<Skills> getSkillsForCandidate(Long candidateId) {
+        List<Skills> skillsForCandidate = []
+
+        sql.eachRow("SELECT name FROM skills " +
+                "INNER JOIN candidate_skills ON skills.skill_id = candidate_skills.skill_id " +
+                "WHERE candidate_skills.candidate_id = ?", [candidateId]) { row ->
+            skillsForCandidate.add(Skills.valueOf(row.name))
+        }
+
+        return skillsForCandidate
+    }
+
+    void update(String cpf, Candidate candidate) {
         try {
             sql.execute("UPDATE candidates SET name = ?, email = ?, cpf = ?, age = ?, state = ?, cep = ?, description = ? WHERE cpf = ?", [
                     candidate.name,
@@ -58,7 +74,7 @@ class CandidateDAO {
             ])
 
             def id = sql.firstRow("SELECT id FROM candidates WHERE cpf = ?", [cpf]).id
-            sql.execute("DELETE FROM candidate_companies WHERE candidate_id = ?", [id])
+            sql.execute("DELETE FROM candidate_skills WHERE candidate_id = ?", [id])
 
             candidate.skillsList.each { skill ->
                 def skillId = sql.firstRow("SELECT skill_id FROM skills WHERE name = ?", [skill])?.skill_id
@@ -66,7 +82,7 @@ class CandidateDAO {
                     sql.execute("INSERT INTO skills (name) VALUES (?)", [skill])
                     skillId = sql.firstRow("SELECT skill_id FROM skills WHERE name = ?", [skill]).skill_id
                 }
-                sql.execute("INSERT INTO candidate_companies (candidate_id, skill_id) VALUES (?, ?)", [id, skillId])
+                sql.execute("INSERT INTO candidate_skills (candidate_id, skill_id) VALUES (?, ?)", [id, skillId])
             }
 
             println("Candidato e suas competências atualizados com sucesso!")
@@ -75,24 +91,21 @@ class CandidateDAO {
             println("Erro ao atualizar candidato: ${e.message}")
         }
     }
-    void Delete(String cpf) {
+    void delete(String cpf) {
         try {
-            def id = sql.firstRow("SELECT id FROM candidates WHERE cpf = ?", [cpf]).id
+            def idRow = sql.firstRow("SELECT id FROM candidates WHERE cpf = ?", [cpf])
+            if (idRow) {
+                def id = idRow.id
 
-            sql.execute("DELETE FROM candidate_companies WHERE candidate_id = ?", [id])
+                sql.execute("DELETE FROM candidate_skills WHERE candidate_id = ?", [id])
+                sql.execute("DELETE FROM candidates WHERE cpf = ?", [cpf])
 
-            sql.execute("DELETE FROM candidates WHERE cpf = ?", [cpf])
-
-            println("Candidato removido com sucesso!")
-
+                println("Candidato removido com sucesso!")
+            } else {
+                println "Candidato com CPF $cpf não encontrado."
+            }
         } catch (Exception e) {
             println("Erro ao remover candidato: ${e.message}")
         }
     }
 }
-
-
-
-
-
-
